[FEATURE] Add Username and creation date to redirects
This feature adds the information when a redirect was
created and by whom.

It is added as a custom form element displaying
information: avatar, realName, username, creation date.

The information is only displayed for stored
record. In case the user can not be determined,
this is displayed accordingly.

Resolves: #107755
Releases: main
Change-Id: I96b933510a00d69125bd93c98e0cd0d66a639546
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91295
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oli Bartsch <bo@cedev.de>